feat: 모달->페이지 교체
