+++ src/features/feed/index.ts
@@ -3,11 +3,13 @@
 
 // feed model
 export { useFeedData } from './model/useFeedData';
+export { useDetailFeedData } from './model/useDetailFeedData';
 
 // feed UI
+export { default as FeedBottomTab } from './ui/FeedBottomTab';
 export { default as FeedCard } from './ui/FeedCard';
-export { default as FeedDetailModal } from './ui/FeedDetailModal';
 export { default as RecipeDetailSection } from './ui/RecipeDetailSection';
 export { default as RecipeSteps } from './ui/RecipeSteps';
 // feed/header UI
 export { default as MyFeedExtension } from './ui/header/MyFeedExtension';
+export { default as HeaderRightContent } from './ui/header/HeaderRightContent';
