+++ src/pages/feed/index.ts
@@ -1,1 +1,2 @@
 export { default as Feed } from './ui/Feed';
+export { default as FeedDetail } from './ui/FeedDetail';
