+++ src/pages/recipe/index.ts
@@ -5,4 +5,5 @@ export {
   RecipeCreateExtension,
   MyRecipeTopSection,
   RecipeCreateTopSection,
+  RecipeDetail,
 } from './ui';
