+++ src/pages/recipe/ui/index.ts
@@ -4,3 +4,4 @@ export { default as MyRecipeExtension } from './MyRecipeExtension';
 export { default as RecipeCreateExtension } from './RecipeCreateExtension';
 export { default as MyRecipeTopSection } from './MyRecipeTopSection';
 export { default as RecipeCreateTopSection } from './RecipeCreateTopSection';
+export { default as RecipeDetail } from './RecipeDetail';
